Move the concurrent.go file location
