+++ database/concurrent.go
@@ -1,4 +1,4 @@
-package util
+package database
 
 import (
 	"sort"
